refactor(git): move unit tests
